# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw Context Engine API
  description: |
    OpenClaw 上下文引擎中间层 API —— 为 AI Agent 提供持久记忆和结构化知识检索。
    
    ## 架构
    - **localMem**: SQLite 记忆存储（2-state 模型：tentative / kept）
    - **LLMWiki**: 结构化知识页面编译和搜索
    
    ## 认证
    配置 `OPENCLAW_API_SECRET` 后，所有请求需要 Bearer Token。
    Unix Socket 连接自动注入 Token。
    
    ## 速率限制
    默认 100 请求/分钟，可通过环境变量配置。
  version: 1.0.0
  contact:
    name: OpenClaw Team
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8901
    description: 本地 HTTP 服务
  - url: http://localhost
    description: Unix Socket 代理（通过 /tmp/openclaw-engine.sock）
security:
  - bearerAuth: []
...
paths:
...
  /api/benchmarks/run:
    post:
      tags: [Benchmark]
      summary: 执行基准测试
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                suite_name:
                  type: string
      responses:
        '200':
          description: 测试执行成功
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API 认证密钥（OPENCLAW_API_SECRET）
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: object
  responses:
...
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: INTERNAL_ERROR
            message: "An unexpected error occurred"
...